Level 3!
Added: third level.
Bugfix/Change: a new, way more accurate, implementation of bouncing that
no more selects the first block (as stored in memory) that would be
hitten by the ball, but the one that the user is expecting the ball
would hit. Fixes the bug where sometimes a ball hitting a block next to
a spike would cause the level to reset and vice-versa.